Refactor Dashboard: Convert chart and polls into presentational components, remove flexbox-react, misc style changes
