+++ src/sass/breakpoints/_1440up.scss
@@ -1,0 +1,5 @@
+.pollsPollTitle {
+  span {
+    font-size: 20px;
+  }
+}
