--- src/sass/breakpoints/_base-noflex.scss
@@ -1,206 +1,0 @@
-/****************************************************************************************************
-Base Styles
-****************************************************************************************************/
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: Josefin Sans, sans-serif;
-  font-size: 20px;
-}
-
-ul {
-  list-style: none;
-  padding-left: 0;
-}
-
-button:active,
-button:focus,
-button:hover,
-input:active,
-input:focus,
-input:hover
-{
-  outline: none;
-}
-
-
-
-/****************************************************************************************************
-Component Styles
-****************************************************************************************************/
-
-.sideMargins {
-  margin-left: 100;
-  margin-right: 100
-}
-
-.textInputStyle {
-  margin: 20px;
-  width: 50%;
-  padding: 10px;
-  background-image: url('%PUBLIC_URL%/input-bg1.png');
-  background-repeat: no-repeat;
-  background-position: center bottom;
-  border: none;
-  text-align: center;
-  font-size: 1rem;
-}
-
-.textInputStyle2 {
-  margin: 20;
-  width: 50%;
-  padding: 10;
-  background-image: url('%PUBLIC_URL%/input-bg2.png');
-  background-repeat: no-repeat;
-  background-position: center bottom;
-  border: none;
-  text-align: center;
-  font-size: 1rem;
-}
-
-.headerStyle {
-  padding-bottom: 10;
-  background-image: url('%PUBLIC_URL%/input-bg1.png');
-  background-repeat: no-repeat;
-  background-position: center bottom;
-}
-
-
-
-//****NavBar
-.linksContainerStyle {
-  justify-content: flex-end;
-  align-self: center;
-  min-width: 50%;
-}
-
-.linkStyle {
-  padding: 20;
-  font-family: Josefin Sans;
-  text-transform: uppercase;
-  text-decoration: none;
-  font-size: 24px;
-  color: #2f3135;
-}
-
-.navbarStyle {
-  justify-content: space-between;
-  margin-bottom: 60;
-  padding: 30px 40px;
-  box-shadow: 0px 4px 4px #CCC;
-}
-
-.logoStyle {
-  margin-right: 8%;
-  font-family: Kaushan Script;
-  font-size: 40px;
-  color: #EC0928;
-  text-decoration: none;
-}
-
-
-
-//****PollsContainer
-
-
-
-//****PollsPoll
-.pollsPollStyle {
-  width: 80%;
-  height: 300px;
-  margin: 24px;
-}
-
-.pollsPollTitle {
-  position: relative;
-  top: 0%;
-  left: 100%;
-  max-width: 60%;
-  padding: 12px 24px;
-  background-color: red;
-  transform: translate(-100%, -100%);
-  box-shadow: 4px 4px 4px rgba(10,10,10,0.5);
-
-  span {
-    text-align: right;
-    word-wrap: break-word;
-    text-transform: uppercase;
-  }
-}
-
-.pollImage {
-  position: absolute;
-  display: block;
-  max-width: 60%;
-  max-height: 280px;
-  min-width: 34%;
-  box-shadow: 4px 4px 4px #CCC;
-  background-color: white;
-
-  img {
-    display: block;
-    max-width: 100%;
-    max-height: 220px;
-  }
-}
-
-.pollImageLeft {
-  left: 10%;
-}
-
-.pollImageRight {
-  position: relative;
-  top: 300px;
-  left: 100%;
-  transform: translate(-100%, -100%);
-}
-
-
-
-//****CreatePoll
-.optionTitleStyle {
-  margin-bottom: 0;
-}
-
-.buttonContainerStyle {
-  background-image: url('%PUBLIC_URL%/btn-bg-grey.png');
-  background-repeat: no-repeat;
-  background-size: contain;
-  background-position: center;
-}
-
-.buttonInputStyle {
-  background: none;
-  margin: 10px;
-  padding: 10px;
-  border: none;
-  text-align: center;
-  font-size: 1rem;
-  color: gray;
-}
-
-
-
-//****Poll
-.pollContainerStyle {
-  margin-top: 0;
-  margin-bottom: 60
-}
-
-.pollOptionContainerStyle {
-  width: 40%;
-  text-align: -webkit-center
-}
-
-.pollOptionImageStyle {
-  display: block;
-  margin-bottom: 20;
-  max-height: 400px;
-  max-width: 90%;
-  width: auto;
-}
-
-.doughnutStyle {
-  margin-bottom: 20
-}
