+++ src/stylesheets/screen.css
@@ -71,7 +71,7 @@ Component Styles
   margin: 20px;
   width: 50%;
   padding: 10px;
-  background-image: url("%PUBLIC_URL%/input-bg1.png");
+  background-image: url("/public/input-bg1.png");
   background-repeat: no-repeat;
   background-position: center bottom;
   border: none;
@@ -84,7 +84,7 @@ Component Styles
   margin: 20;
   width: 50%;
   padding: 10;
-  background-image: url("%PUBLIC_URL%/input-bg2.png");
+  background-image: url("/public/input-bg2.png");
   background-repeat: no-repeat;
   background-position: center bottom;
   border: none;
@@ -92,18 +92,10 @@ Component Styles
   font-size: 1rem;
 }
 
-/* line 62, ../sass/breakpoints/_base.scss */
-.headerStyle {
-  padding-bottom: 10;
-  background-image: url("%PUBLIC_URL%/input-bg1.png");
-  background-repeat: no-repeat;
-  background-position: center bottom;
-}
-
 /**************************************************
-Nav Bar
+  Nav Bar
 **************************************************/
-/* line 75, ../sass/breakpoints/_base.scss */
+/* line 68, ../sass/breakpoints/_base.scss */
 .navbarStyle {
   display: flex;
   flex-wrap: wrap;
@@ -113,7 +105,7 @@ Nav Bar
   box-shadow: 0px 4px 4px #CCC;
 }
 
-/* line 84, ../sass/breakpoints/_base.scss */
+/* line 77, ../sass/breakpoints/_base.scss */
 .logoStyle {
   margin-right: 8%;
   font-family: Kaushan Script;
@@ -122,7 +114,7 @@ Nav Bar
   text-decoration: none;
 }
 
-/* line 92, ../sass/breakpoints/_base.scss */
+/* line 85, ../sass/breakpoints/_base.scss */
 .linksContainerStyle {
   display: flex;
   justify-content: flex-end;
@@ -130,7 +122,7 @@ Nav Bar
   min-width: 50%;
 }
 
-/* line 99, ../sass/breakpoints/_base.scss */
+/* line 92, ../sass/breakpoints/_base.scss */
 .linkStyle {
   padding: 20px;
   font-family: Josefin Sans;
@@ -142,9 +134,27 @@ Nav Bar
 }
 
 /**************************************************
+  Dashboard
+**************************************************/
+/* line 108, ../sass/breakpoints/_base.scss */
+.dashboard__header {
+  padding-bottom: 10;
+  background-image: url("/public/input-bg1.png");
+  background-repeat: no-repeat;
+  background-position: center bottom;
+}
+
+/* line 115, ../sass/breakpoints/_base.scss */
+.dashboard__contents {
+  display: flex;
+  justify-content: space-around;
+  align-content: center;
+}
+
+/**************************************************
 Polls List
 **************************************************/
-/* line 116, ../sass/breakpoints/_base.scss */
+/* line 128, ../sass/breakpoints/_base.scss */
 .pollsList {
   display: flex;
   flex-wrap: wrap;
@@ -156,43 +166,43 @@ Polls List
   margin-right: auto;
 }
 
-/* line 130, ../sass/breakpoints/_base.scss */
+/* line 142, ../sass/breakpoints/_base.scss */
 .pollsPollStyle {
   width: 80%;
   padding: 0% 0% 12% 0%;
 }
-/* line 134, ../sass/breakpoints/_base.scss */
+/* line 146, ../sass/breakpoints/_base.scss */
 .pollsPollStyle .pollsPollTitle {
   animation: easeUp 0.5s;
   animation-fill-mode: forwards;
 }
-/* line 138, ../sass/breakpoints/_base.scss */
+/* line 150, ../sass/breakpoints/_base.scss */
 .pollsPollStyle .pollImageLeft img {
   animation: easeRight 0.5s;
   animation-fill-mode: forwards;
 }
-/* line 142, ../sass/breakpoints/_base.scss */
+/* line 154, ../sass/breakpoints/_base.scss */
 .pollsPollStyle .pollImageRight img {
   animation: easeLeft 0.5s;
   animation-fill-mode: forwards;
 }
-/* line 148, ../sass/breakpoints/_base.scss */
+/* line 160, ../sass/breakpoints/_base.scss */
 .pollsPollStyle:hover .pollsPollTitle {
   animation: easeDown 1s ease-out;
   animation-fill-mode: forwards;
 }
-/* line 152, ../sass/breakpoints/_base.scss */
+/* line 164, ../sass/breakpoints/_base.scss */
 .pollsPollStyle:hover .pollImageLeft img {
   animation: easeLeft 1s ease-out;
   animation-fill-mode: forwards;
 }
-/* line 156, ../sass/breakpoints/_base.scss */
+/* line 168, ../sass/breakpoints/_base.scss */
 .pollsPollStyle:hover .pollImageRight img {
   animation: easeRight 1s ease-out;
   animation-fill-mode: forwards;
 }
 
-/* line 163, ../sass/breakpoints/_base.scss */
+/* line 175, ../sass/breakpoints/_base.scss */
 .pollsPollTitle {
   position: relative;
   width: 50%;
@@ -203,7 +213,7 @@ Polls List
   border: 0.5px solid lightgray;
   border-left: 4px solid #ff2955;
 }
-/* line 173, ../sass/breakpoints/_base.scss */
+/* line 185, ../sass/breakpoints/_base.scss */
 .pollsPollTitle span {
   text-align: right;
   word-wrap: break-word;
@@ -211,7 +221,7 @@ Polls List
   font-weight: bold;
 }
 
-/* line 181, ../sass/breakpoints/_base.scss */
+/* line 193, ../sass/breakpoints/_base.scss */
 .pollImageFlex {
   display: flex;
   flex-wrap: nowrap;
@@ -221,7 +231,7 @@ Polls List
   height: 41.37931vw;
 }
 
-/* line 190, ../sass/breakpoints/_base.scss */
+/* line 202, ../sass/breakpoints/_base.scss */
 .pollImage {
   display: block;
   max-height: 280px;
@@ -230,12 +240,12 @@ Polls List
   border-radius: 2px;
   background-color: white;
 }
-/* line 198, ../sass/breakpoints/_base.scss */
+/* line 210, ../sass/breakpoints/_base.scss */
 .pollImage img {
   max-width: 100%;
 }
 
-/* line 203, ../sass/breakpoints/_base.scss */
+/* line 215, ../sass/breakpoints/_base.scss */
 .pollImageLeft {
   min-width: 50%;
   min-height: 50%;
@@ -245,7 +255,7 @@ Polls List
   margin-right: -10%;
 }
 
-/* line 212, ../sass/breakpoints/_base.scss */
+/* line 224, ../sass/breakpoints/_base.scss */
 .pollImageRight {
   max-width: 80%;
   max-height: 220px;
@@ -253,20 +263,20 @@ Polls List
   margin-left: -10%;
 }
 
-/* line 223, ../sass/breakpoints/_base.scss */
+/* line 235, ../sass/breakpoints/_base.scss */
 .optionTitleStyle {
   margin-bottom: 0;
 }
 
-/* line 227, ../sass/breakpoints/_base.scss */
+/* line 239, ../sass/breakpoints/_base.scss */
 .buttonContainerStyle {
-  background-image: url("%PUBLIC_URL%/btn-bg-grey.png");
+  background-image: url("/public/btn-bg-grey.png");
   background-repeat: no-repeat;
   background-size: contain;
   background-position: center;
 }
 
-/* line 234, ../sass/breakpoints/_base.scss */
+/* line 246, ../sass/breakpoints/_base.scss */
 .buttonInputStyle {
   background: none;
   margin: 10px;
@@ -277,19 +287,19 @@ Polls List
   color: gray;
 }
 
-/* line 247, ../sass/breakpoints/_base.scss */
+/* line 259, ../sass/breakpoints/_base.scss */
 .pollContainerStyle {
   margin-top: 0;
   margin-bottom: 60;
 }
 
-/* line 252, ../sass/breakpoints/_base.scss */
+/* line 264, ../sass/breakpoints/_base.scss */
 .pollOptionContainerStyle {
   width: 40%;
   text-align: -webkit-center;
 }
 
-/* line 257, ../sass/breakpoints/_base.scss */
+/* line 269, ../sass/breakpoints/_base.scss */
 .pollOptionImageStyle {
   display: block;
   margin-bottom: 20;
@@ -298,7 +308,7 @@ Polls List
   width: auto;
 }
 
-/* line 265, ../sass/breakpoints/_base.scss */
+/* line 277, ../sass/breakpoints/_base.scss */
 .doughnutStyle {
   margin-bottom: 20;
 }
